chore: upgrade to Tauri 2.0 (alpha)
