+++ package.json
@@ -23,16 +23,16 @@
     "@solid-primitives/context": "^0.2.1",
     "@solid-primitives/scheduled": "^1.4.1",
     "@tailwindcss/typography": "^0.5.10",
-    "@tauri-apps/cli": "1.4.0",
-    "@types/node": "^20.6.0",
-    "@typescript-eslint/eslint-plugin": "^6.6.0",
-    "@typescript-eslint/parser": "^6.6.0",
-    "autoprefixer": "^10.4.15",
+    "@tauri-apps/cli": "1.5.2",
+    "@types/node": "^20.8.5",
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "autoprefixer": "^10.4.16",
     "conventional-changelog-conventionalcommits": "^7.0.2",
-    "daisyui": "^3.7.3",
-    "eslint": "^8.49.0",
+    "daisyui": "^3.9.2",
+    "eslint": "^8.51.0",
     "eslint-plugin-tsdoc": "^0.2.17",
-    "postcss": "^8.4.29",
+    "postcss": "^8.4.31",
     "postcss-import": "^15.1.0",
     "postcss-nesting": "^12.0.1",
     "prettier": "^3.0.3",
@@ -40,13 +40,18 @@
     "solid-icons": "^1.0.12",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.2.2",
-    "vite": "^4.4.9",
-    "vite-plugin-solid": "^2.7.0"
+    "vite": "^4.4.11",
+    "vite-plugin-solid": "^2.7.1"
   },
   "dependencies": {
-    "@tauri-apps/api": "1.4.0",
+    "@tauri-apps/api": "2.0.0-alpha.8",
+    "@tauri-apps/plugin-app": "2.0.0-alpha.1",
+    "@tauri-apps/plugin-dialog": "2.0.0-alpha.1",
+    "@tauri-apps/plugin-fs": "2.0.0-alpha.1",
+    "@tauri-apps/plugin-store": "2.0.0-alpha.1",
+    "@tauri-apps/plugin-window": "2.0.0-alpha.1",
     "minisearch": "^6.1.0",
-    "solid-js": "^1.7.11"
+    "solid-js": "^1.8.1"
   },
   "prettier": {
     "bracketSameLine": true,
