+++ src-tauri/Cargo.toml
@@ -12,27 +12,19 @@ rust-version = "1.57"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.4", features = [] }
+tauri-build = { version = "2.0.0-alpha.9", features = [] }
 
 [dependencies]
-tauri = { version = "1.4", features = [
-    "dialog-open",
-    "fs-read-dir",
-    "protocol-asset",
-    "updater",
-    "window-set-title",
-] }
-log = "0.4.20"
 chrono = "0.4"
+log = "0.4.20"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 
-# [dependencies.fern]
-# default-features = false
-# version = "0.6"
-# features = ["colored"]
-
-[dependencies.tauri-plugin-store]
-git = "https://github.com/tauri-apps/plugins-workspace"
-branch = "dev"
+[dependencies.tauri]
+version = "2.0.0-alpha"
+features = [
+    "protocol-asset",
+]
 
 [dependencies.dfraw_json_parser]
 # git = "https://github.com/nwesterhausen/dfraw_json_parser"
@@ -43,11 +35,35 @@ path = "../../dfraw_json_parser"
 
 [dependencies.tauri-plugin-log]
 git = "https://github.com/tauri-apps/plugins-workspace"
-branch = "dev"
+branch = "v2"
 
-[dependencies.tauri-plugin-window-state]
+# Todo: un-remove this when it's fixed
+# Current issue is from the window-state plugin library
+#1 |     TauriApi(#[from] tauri::api::Error),
+#  |                             ^^^ could not find `api` in `tauri`
+# [dependencies.tauri-plugin-window-state]
+# git = "https://github.com/tauri-apps/plugins-workspace"
+# branch = "v2"
+
+[dependencies.tauri-plugin-window]
 git = "https://github.com/tauri-apps/plugins-workspace"
-branch = "dev"
+branch = "v2"
+
+[dependencies.tauri-plugin-fs]
+git = "https://github.com/tauri-apps/plugins-workspace"
+branch = "v2"
+
+[dependencies.tauri-plugin-app]
+git = "https://github.com/tauri-apps/plugins-workspace"
+branch = "v2"
+
+[dependencies.tauri-plugin-store]
+git = "https://github.com/tauri-apps/plugins-workspace"
+branch = "v2"
+
+[dependencies.tauri-plugin-dialog]
+git = "https://github.com/tauri-apps/plugins-workspace"
+branch = "v2"
 
 [features]
 # by default Tauri runs in production mode
@@ -63,3 +79,7 @@ codegen-units = 1
 lto = true
 opt-level = "s"
 strip = true      # Automatically strip symbols from the binary.
+
+[lib]
+crate-type = ["staticlib", "cdylib", "rlib"]
+name = "app_lib"
